First version
I strived for a functional style, but it still feels a bit imperative
(indices? Seriously?)

The only further optimization that I can think of is remembering current
best and terminating the cases where we won't be able to improve it.
Unfortunately, that will kill all the FP for sure. Version three will be
written in State monad.