Second version, fixed
Now it not only carries "best" solution around, but actually uses it to
cut branches off. I also re-wrote the thing in tail-recursive manner.

After a good night's sleep I realised that I don't actually need State
monad for this - it's not that bad as it is. But I might try to rewrite
it with ByteStrings or something else that allows to iterate over the
string from both ends; that should improve memory consumption by
a factor of two as I won't need `reverse string` anymore.